+++ v3/crates/jsonapi/src/process_response.rs
@@ -1,4 +1,8 @@
+use super::helpers::get_object_type;
 use super::types::{RelationshipNode, RelationshipTree};
+use crate::catalog::ObjectType;
+use crate::RequestError;
+use jsonapi_library::query::Query;
 use metadata_resolve::Qualified;
 use open_dds::{relationships::RelationshipType, types::CustomTypeName};
 use std::collections::BTreeMap;
@@ -24,8 +28,10 @@ fn to_resource(
     rowset: ndc_models::RowSet,
     type_name: &Qualified<CustomTypeName>,
     relationship_tree: &RelationshipTree,
+    query: &Query,
+    object_types: &BTreeMap<Qualified<CustomTypeName>, ObjectType>,
     collect_relationships: &mut Vec<jsonapi_library::model::Resource>,
-) -> Vec<jsonapi_library::model::Resource> {
+) -> Result<Vec<jsonapi_library::model::Resource>, RequestError> {
     let mut resources = vec![];
     if let Some(rows) = rowset.rows {
         for row in rows {
@@ -36,12 +42,14 @@ fn to_resource(
                 collect_relationships,
                 resource_id,
                 type_name,
+                query,
+                object_types,
                 row.into_iter().map(|(k, v)| (k.to_string(), v.0)),
-            );
+            )?;
             resources.push(resource);
         }
     }
-    resources
+    Ok(resources)
 }
 
 fn render_type_name(type_name: &Qualified<CustomTypeName>) -> String {
@@ -54,10 +62,13 @@ fn row_to_resource(
     collect_relationships: &mut Vec<jsonapi_library::model::Resource>,
     resource_id: i32,
     row_type: &Qualified<CustomTypeName>,
+    query: &Query,
+    object_types: &BTreeMap<Qualified<CustomTypeName>, ObjectType>,
     row: impl Iterator<Item = (String, serde_json::Value)>,
-) -> jsonapi_library::model::Resource {
+) -> Result<jsonapi_library::model::Resource, RequestError> {
     let mut attributes = BTreeMap::new();
     let mut relationships = BTreeMap::new();
+
     for (key, mut value) in row {
         // Check if the key is a relationship
         if let Some(relationship_node) = relationship_tree.relationships.get(key.as_str()) {
@@ -89,8 +100,10 @@ fn row_to_resource(
                                 relationship_id,
                                 object_type,
                                 *is_command_relationship,
+                                query,
+                                object_types,
                                 object_row_value,
-                            );
+                            )?;
                             jsonapi_library::model::IdentifierData::Single(resource_identifier)
                         } else {
                             jsonapi_library::model::IdentifierData::None
@@ -112,8 +125,10 @@ fn row_to_resource(
                                 relationship_id,
                                 object_type,
                                 *is_command_relationship,
+                                query,
+                                object_types,
                                 object_row_value.take(),
-                            );
+                            )?;
                             resource_identifiers.push(resource_identifier);
                         }
                         jsonapi_library::model::IdentifierData::Multiple(resource_identifiers)
@@ -126,12 +141,19 @@ fn row_to_resource(
             };
             relationships.insert(key.to_string(), relationship);
         } else {
-            attributes.insert(key.to_string(), value);
+            let object_type =
+                get_object_type(object_types, row_type).map_err(RequestError::ParseError)?;
+            let identifier = open_dds::identifier::Identifier::new(key.as_str()).unwrap();
+            let field_name = open_dds::types::FieldName::new(identifier);
+
+            if object_type.type_fields.contains_key(&field_name) {
+                attributes.insert(key.to_string(), value);
+            }
         }
     }
 
     let rendered_type_name = render_type_name(row_type);
-    jsonapi_library::api::Resource {
+    Ok(jsonapi_library::api::Resource {
         _type: rendered_type_name,
         id: resource_id.to_string(),
         attributes,
@@ -142,7 +164,7 @@ fn row_to_resource(
         } else {
             Some(relationships)
         },
-    }
+    })
 }
 
 fn collect_relationship_value(
@@ -152,8 +174,11 @@ fn collect_relationship_value(
     resource_id: i32,
     row_type: &Qualified<CustomTypeName>,
     is_command_relationship: bool,
+    query: &Query,
+    object_types: &BTreeMap<Qualified<CustomTypeName>, ObjectType>,
+
     mut value: serde_json::Value,
-) {
+) -> Result<(), RequestError> {
     if is_command_relationship {
         // If this is a command relationship, we need to extract the value from the 'FUNCTION_IR_VALUE_COLUMN_NAME' key
         // We are ignoring the other keys
@@ -172,17 +197,23 @@ fn collect_relationship_value(
         collect_relationships,
         resource_id,
         row_type,
+        query,
+        object_types,
         row_object.into_iter(),
-    );
+    )?;
     // collect this relationship resource
     collect_relationships.push(relationship_resource);
+
+    Ok(())
 }
 
 pub fn process_result(
     rowsets: Vec<ndc_models::RowSet>,
     root_type_name: &Qualified<CustomTypeName>,
     relationship_tree: &RelationshipTree,
-) -> jsonapi_library::api::DocumentData {
+    query: &Query,
+    object_types: &BTreeMap<Qualified<CustomTypeName>, ObjectType>,
+) -> Result<jsonapi_library::api::DocumentData, RequestError> {
     let mut unique_id = 1;
 
     let mut resources = vec![];
@@ -193,8 +224,10 @@ pub fn process_result(
             first_rowset,
             root_type_name,
             relationship_tree,
+            query,
+            object_types,
             &mut collect_relationships,
-        ));
+        )?);
     }
 
     let included = if collect_relationships.is_empty() {
@@ -203,11 +236,11 @@ pub fn process_result(
         Some(collect_relationships)
     };
 
-    jsonapi_library::api::DocumentData {
+    Ok(jsonapi_library::api::DocumentData {
         data: Some(jsonapi_library::api::PrimaryData::Multiple(resources)),
         included,
         links: None,
         meta: None,
         jsonapi: None,
-    }
+    })
 }